fix(payments): mentorship never delivered, and the poller never ran
Two bugs the integration suite could only surface one per twenty-five
minute run, because nextest stops a shard at its first failure. Running it
locally found both in minutes.

Paying for a mentorship session took the money and delivered nothing.
fulfilment set updated_at on mentorship_sessions, a column that table has
never had, so every mentorship payment succeeded and then failed to
deliver: session left pending, row handed to the sweep to retry forever.
That is the paid-but-undelivered count the admin dashboard exists to show.

EXTRACT(EPOCH ...) is numeric in Postgres, not float8, and three places
decoded it straight into an f64: the status endpoint, the background
poller and reconciliation. All three failed at runtime — which means the
poller, the thing that makes a payment confirm whether or not the browser
comes back, did not run at all.

Also: --no-fail-fast in CI, so one run reports every failure instead of
four; and four tests funded through the ledger, because the withdraw
endpoint checks the balance before the destination and the KYC status, so
an unfunded account is refused for a reason the test was not about.

+++ src/routes/payments.rs
@@ -316,7 +316,7 @@ pub async fn status(
     let payment: Row = sqlx::query_as(
         "SELECT payer_id, provider, status, merchant_reference, provider_reference,
                 provider_session_id, fulfilled_at,
-                EXTRACT(EPOCH FROM (NOW() - last_checked_at)) AS since_last_ask
+                EXTRACT(EPOCH FROM (NOW() - last_checked_at))::float8 AS since_last_ask
            FROM payments WHERE id = $1",
     )
     .bind(id)
